Fix circular question template circular dependency
This was a bug that has been bugging me. For weeks actually. I've not
had the time to sit down and focus on it, so instead I was just putting
up with failing tests.

At fault, there were three functions, all to do with constructing
questions in english. They depended on eniglish-template, which is not
allowed in the english transformations module because it creates a
circular dependency.

To fix, I've moved them (perhaps temporarily) to the
compose-english-phrases module. Which is where they belong anyway I
think.